requirements/base.txt
Django==5.1.13
djangorestframework==3.16.1
pytest==8.4.2
pytest-django==4.9.0
psycopg==3.2.10
# django-cron==0.6.0
redis==6.4.0
django-redis==6.0.0
celery==5.5.3
requests==2.32.5
drf-yasg==1.21.11
